add snapshot test + move other tests to __tests__ folder
