+++ src/shared/components/Header/__tests__/Header.test.tsx
@@ -1,6 +1,6 @@
-import React = require('react');
+import * as React from 'react';
 import { shallow, mount } from 'enzyme';
-import Header, { HeaderProps } from '.';
+import Header, { HeaderProps } from '..';
 
 const links: React.ReactNode[] = [
   <p>A p tag</p>,
